move fruits to new utils ns
