+++ src/com/untamedears/utils/Fruits.java
@@ -1,4 +1,4 @@
-package com.untamedears.realisticbiomes.persist;
+package com.untamedears.utils;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -7,6 +7,8 @@ import org.bukkit.Material;
 import org.bukkit.block.Block;
 import org.bukkit.block.BlockFace;
 import org.bukkit.util.Vector;
+
+import com.untamedears.realisticbiomes.persist.BlockGrower;
 
 /**
  * Fruit related helper functions. This could be optimized: convert this
